LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_1-expected.txt


FAIL Removing iframe from document removes it from history assert_equals: expected 2 but got 1
